Extract common permission related code
