+++ gravatar-quickeditor/src/main/java/com/gravatar/quickeditor/ui/components/AvatarsSection.kt
@@ -1,11 +1,8 @@
 package com.gravatar.quickeditor.ui.components
 
 import android.Manifest
-import android.content.Context
-import android.content.pm.PackageManager
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.StringRes
@@ -18,15 +15,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import com.gravatar.quickeditor.QuickEditorFileProvider
 import com.gravatar.quickeditor.R
 import com.gravatar.quickeditor.ui.avatarpicker.AvatarUi
 import com.gravatar.quickeditor.ui.avatarpicker.AvatarsSectionUiState
 import com.gravatar.quickeditor.ui.editor.AvatarPickerContentLayout
-import com.gravatar.quickeditor.ui.oauth.findComponentActivity
+import com.gravatar.quickeditor.ui.hasCameraPermissionInManifest
 import com.gravatar.quickeditor.ui.openAppPermissionSettings
+import com.gravatar.quickeditor.ui.withPermission
 import com.gravatar.restapi.models.Avatar
 import com.gravatar.ui.GravatarTheme
 import java.net.URI
@@ -67,13 +63,20 @@ internal fun AvatarsSection(
     }
 
     val permissionAwareTakePhotoCallback = {
-        context.withCameraPermission(
-            cameraPermissionLauncher = cameraPermissionLauncher,
-            onShowRationale = { cameraPermissionDialogRationaleVisible = true },
-            grantedCallback = {
-                takePhotoCallback()
-            },
-        )
+        if (context.hasCameraPermissionInManifest()) {
+            context.withPermission(
+                permission = Manifest.permission.CAMERA,
+                onRequestPermission = {
+                    cameraPermissionLauncher.launch(it)
+                },
+                onShowRationale = { cameraPermissionDialogRationaleVisible = true },
+                grantedCallback = {
+                    takePhotoCallback()
+                },
+            )
+        } else {
+            takePhotoCallback()
+        }
     }
 
     when (state.avatarPickerContentLayout) {
@@ -113,35 +116,6 @@ internal fun AvatarsSection(
             context.openAppPermissionSettings()
         },
     )
-}
-
-internal fun Context.withCameraPermission(
-    cameraPermissionLauncher: ActivityResultLauncher<String>,
-    onShowRationale: () -> Unit = {},
-    grantedCallback: () -> Unit,
-) {
-    if (hasCameraPermissionInManifest()) {
-        val activity = findComponentActivity()
-        when {
-            ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.CAMERA,
-            ) == PackageManager.PERMISSION_GRANTED -> {
-                grantedCallback()
-            }
-
-            activity != null &&
-                ActivityCompat.shouldShowRequestPermissionRationale(activity, Manifest.permission.CAMERA) -> {
-                onShowRationale()
-            }
-
-            else -> {
-                cameraPermissionLauncher.launch(Manifest.permission.CAMERA)
-            }
-        }
-    } else {
-        grantedCallback()
-    }
 }
 
 internal val AvatarsSectionUiState.titleRes: Int
@@ -227,11 +201,4 @@ private fun AvatarSectionEmptyPreview() {
             onLocalImageSelected = { },
         )
     }
-}
-
-private fun Context.hasCameraPermissionInManifest(): Boolean {
-    val packageInfo = packageManager.getPackageInfo(packageName, PackageManager.GET_PERMISSIONS)
-    val permissions = packageInfo.requestedPermissions
-
-    return permissions?.any { perm -> perm == Manifest.permission.CAMERA } ?: false
 }
